+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 2.0.25
 ++++++
+* Adding paging support for storage list commands.
 * Minor fixes.
 
 2.0.24
@@ -15,7 +16,6 @@ Release History
 * Added 'snapshot' column to table output for blob list/show.
 * Fixed bugs with various parameters that needed to be parsed as ints, added test coverage.
 * Small fix with test, `storage blob service-properties show`: "hourMetrics.enabled" defaults to false.
-* Minor fixes
 
 2.0.23
 ++++++